+++ modules/azure/main.tf
@@ -42,10 +42,11 @@ module "custom_data" {
   location  = var.location
   api_key   = var.api_key != null ? var.api_key : "@Microsoft.KeyVault(SecretUri=${local.api_key_uri})"
   site      = var.site
-  client_id = module.managed_identity.identity.client_id
+  client_id = length(module.managed_identity) == 0 ? var.scanner_identity.client_id : module.managed_identity[0].identity.client_id
 }
 
 module "managed_identity" {
+  count               = var.scanner_identity == null ? 1 : 0
   source              = "./managed-identity"
   resource_group_name = module.resource_group.resource_group.name
   resource_group_id   = module.resource_group.resource_group.id
@@ -63,6 +64,6 @@ module "virtual_machine" {
   admin_ssh_key          = var.admin_ssh_key
   custom_data            = module.custom_data.install_sh
   subnet_id              = module.virtual_network.subnet.id
-  user_assigned_identity = module.managed_identity.identity.id
+  user_assigned_identity = length(module.managed_identity) == 0 ? var.scanner_identity.id : module.managed_identity[0].identity.id
   tags                   = var.tags
 }